fix(web): abbreviate month names in experience date ranges
December -> Dec, November -> Nov, etc. Only affects formatMonthYear
(used for Experience date ranges); formatDate (Writing post dates)
keeps full month names.

+++ apps/web/lib/format.ts
@@ -1,4 +1,7 @@
-const MONTH_YEAR: Intl.DateTimeFormatOptions = { month: "long", year: "numeric" };
+const MONTH_YEAR: Intl.DateTimeFormatOptions = {
+  month: "short",
+  year: "numeric",
+};
 
 export function formatDate(iso: string): string {
   return new Date(iso).toLocaleDateString("en-US", {
